+++ src/com/emmaobo/expensecalculator/interfaces/Writer.java
@@ -1,8 +1,8 @@
 package com.emmaobo.expensecalculator.interfaces;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 
 public interface Writer {
-
-    void writeList(HashMap<String, Double> list);
+    void writeList(HashMap<String, BigDecimal> list);
 }
